Customizing user authentication

+++ app/Http/Controllers/PostController.php
@@ -5,8 +5,9 @@ namespace App\Http\Controllers;
 use App\Like;
 use App\Post;
 use App\Tag;
+use Auth;
+use Gate;
 use Illuminate\Http\Request;
-
 use App\Http\Requests;
 
 class PostController extends Controller
@@ -17,6 +18,9 @@ class PostController extends Controller
     }
 
     public function getAdminIndex(){
+        if (!Auth::check()){
+            return redirect()->back();
+        }
         $posts = Post::orderBy('title', 'asc')->get();
         return view('admin.index', ['posts' => $posts]);
     }
@@ -34,12 +38,21 @@ class PostController extends Controller
     }
 
     public function getAdminCreate(){
+        if (!Auth::check()){
+            return redirect()->back();
+        }
         $tags = Tag::all();
         return view('admin.create', ['tags' => $tags]);
     }
 
     public function getAdminEdit($id){
+        if (!Auth::check()){
+            return redirect()->back();
+        }
         $post = Post::find($id);
+        if (Gate::denies('manipulate-post', $post)){
+            return redirect()->back();
+        }
         $tags = Tag::all();
         return view('admin.edit', ['post' => $post, 'postId' => $id, 'tags' => $tags]);
     }
@@ -49,22 +62,32 @@ class PostController extends Controller
             'title' => 'required|min:5',
             'content' => 'required|min:10'
         ]);
+        $user = Auth::user();
+        if (!$user){
+            return redirect()->back();
+        }
         $post = new Post([
             'title' => $request->input('title'),
             'content' => $request->input('content')
         ]);
-        $post->save();
+        $user->posts()->save($post);
         $post->tags()->attach($request->input('tags') === null ? [] : $request->input('tags'));
         
         return redirect()->route('admin.index')->with('info', 'Post created, title is: ' . $request->input('title'));
     }
 
     public function postAdminUpdate(Request $request){
+        if (!Auth::check()){
+            return redirect()->back();
+        }
         $this->validate($request, [
             'title' => 'required|min:5',
             'content' => 'required|min:10'
         ]);
         $post = Post::find($request->input('id'));
+        if (Gate::denies('manipulate-post', $post)){
+            return redirect()->back();
+        }
         $post->title = $request->input('title');
         $post->content = $request->input('content');
         $post->save();
@@ -73,7 +96,13 @@ class PostController extends Controller
     }
 
     public function getAdminDelete($id){
+        if (!Auth::check()){
+            return redirect()->back();
+        }
         $post = Post::find($id);
+        if (Gate::denies('manipulate-post', $post)){
+            return redirect()->back();
+        }
         $post->likes()->delete();
         $post->tags()->detach();
         $post->delete();
